Move generic utility functions from Analysis to Triton IR
This is done to reduce the dependency from Triton IR to TritonGPUAttrDefsIncGen.
