+++ lib/Dialect/Triton/IR/Traits.cpp
@@ -3,8 +3,8 @@
 #include <numeric>
 
 #include "mlir/IR/TypeUtilities.h"
-#include "triton/Analysis/Utility.h"
 #include "triton/Dialect/Triton/IR/Types.h"
+#include "triton/Dialect/Triton/IR/Utility.h"
 #include "triton/Dialect/TritonGPU/IR/Dialect.h"
 
 using namespace mlir;
