+++ lib/Dialect/TritonGPU/Transforms/Coalesce.cpp
@@ -1,6 +1,6 @@
 #include "mlir/Analysis/SliceAnalysis.h"
 #include "triton/Analysis/AxisInfo.h"
-#include "triton/Analysis/Utility.h"
+#include "triton/Dialect/Triton/IR/Utility.h"
 #include "triton/Dialect/TritonGPU/IR/Dialect.h"
 #include "triton/Dialect/TritonGPU/Transforms/Passes.h"
 #include "triton/Dialect/TritonGPU/Transforms/Utility.h"
